Demo static now works for loading the database (stage_4).
Docker/3DCityDBImpExp-DockerContext/Dockerfile now uses a fix SHA1
(because the last commits of 3dCityDBImporterExporter were breaking
things...

The logic (and code) of DemoLoad3DCityDB down-ported to DockerLoad3DCityDB
parent class. For example the generation of the ad-hoc configuration file
or what to check in the log_file belongs to the DockerLoad3DCityDB and
should be hidden away from its users.
As a desired side effect DemoLoad3DCityDB (renamed to DemoLoad3DCityDBTemporal)
got much shorter and much lesser code needs to be repeated in its static
counter-part.

DemoStatic/demo_configuration_static.py: default port is not in collision
with its temporal counterpart.